canActivate option (#714)
I've added an option so that plugins can activate other plugins by adding an extra optional line to the profiles
canActivate: ['plugin1', 'plugin2', ... ]
This will allow that plugin to activate other plugins. But we can curate this when the profiles are made.
This optional. If the other plugin does not exist or the caller doesn't have canActivate in its profile it gets caught.