Rntabs (#893)
This addresses an issue when you rename a folder which has files which are open in tabs.
It would recreate the current opened file because it is current and saved when another file opened. You'd get duplicate files. Also tabs from the old folder would still be visible.
Now the behavior should be: tabs that belong to the folder renamed will be closed and replaced. No file will be focused.
Please test.